asiayo_demo/requirements.txt
asgiref==3.5.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pip-autoremove==0.10.0
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
uWSGI==2.0.20
